+++ sandpack-react/src/hooks/usePreviewProgress.ts
@@ -1,27 +1,7 @@
 import type { WorkerStatusUpdate } from "@codesandbox/nodebox";
 import * as React from "react";
 
-import { useSandpack } from "../..";
-import { css } from "../../styles";
-import { fadeIn } from "../../styles/shared";
-
-const wrapperClassName = css({
-  position: "absolute",
-  left: "$space$5",
-  bottom: "$space$4",
-  zIndex: "$top",
-  color: "$colors$clickable",
-  animation: `${fadeIn} 150ms ease`,
-  fontFamily: "$font$mono",
-  fontSize: ".8em",
-  width: "75%",
-  p: {
-    whiteSpace: "nowrap",
-    margin: 0,
-    textOverflow: "ellipsis",
-    overflow: "hidden",
-  },
-});
+import { useSandpack } from "..";
 
 const mapProgressMessage = (
   originalMessage: WorkerStatusUpdate & { command?: string },
@@ -44,14 +24,13 @@ const mapProgressMessage = (
   }
 };
 
-export const PreviewProgress: React.FC<{
-  clientId?: string;
-}> = ({ clientId }) => {
+export const usePreviewProgress = () => {
   const [isReady, setIsReady] = React.useState(false);
   const [totalDependencies, setTotalDependencies] = React.useState<number>();
   const [loadingMessage, setLoadingMessage] = React.useState<null | string>(
     null
   );
+
   const { listen } = useSandpack();
 
   React.useEffect(() => {
@@ -78,7 +57,7 @@ export const PreviewProgress: React.FC<{
         setIsReady(true);
         clearTimeout(timer);
       }
-    }, clientId);
+    });
 
     return (): void => {
       if (timer) {
@@ -86,13 +65,7 @@ export const PreviewProgress: React.FC<{
       }
       unsubscribe();
     };
-  }, [clientId, isReady, totalDependencies]);
+  }, [isReady, totalDependencies]);
 
-  if (!loadingMessage) return null;
-
-  return (
-    <div className={wrapperClassName.toString()}>
-      <p>{loadingMessage}</p>
-    </div>
-  );
+  return loadingMessage;
 };
